build: update dependencies and refactor hardware port access
- Update `addr2line`, `gimli`, `goblin`, and `heapless` to newer versions in `Cargo.lock`.
- Refactor port I/O calls by removing redundant `crate::` prefixes in hardware port operations.
- Update imports in `petroleum/src/lib.rs` to source port operations from `hardware::ports` instead of `graphics::ports`.
- Simplify PE parsing in `petroleum/src/page_table/pe.rs` by utilizing the `goblin::pe::PE` struct and removing the manual `PeSectionHeader` definition.
