+++ petroleum/Cargo.toml
@@ -22,16 +22,16 @@ linked_list_allocator = "0.10"
 once_cell = { version = "1.19", default-features = false, features = ["alloc", "race"] }
 embedded-graphics = "0.8.1"
 volatile = "0.6"
-heapless = { version = "0.8", default-features = false }
+heapless = { version = "0.9.2", default-features = false }
 bitflags = "2.4"
 conquer-util = { version = "0.3", default-features = false }
 static_assertions = "1.1"
 log = { version = "0.4.28", default-features = false }
 paste = "1.0"
 bitfield = "0.19"
-addr2line = { version = "0.24", default-features = false }
-gimli = { version = "0.30", default-features = false }
-goblin = { version = "0.9", default-features = false }
+addr2line = { version = "0.26.1", default-features = false }
+gimli = { version = "0.33.0", default-features = false }
+goblin = { version = "0.10.5", default-features = false, features = ["pe32", "pe64"] }
 
 [lib]
 name = "petroleum"
