+++ petroleum/src/graphics/mod.rs
@@ -67,8 +67,6 @@ pub mod color;
 pub mod constants;
 pub mod desktop;
 pub mod framebuffer;
-#[macro_use]
-pub mod ports;
 pub mod registers;
 pub mod setup;
 pub mod text;
@@ -78,7 +76,7 @@ pub use constants::*;
 
 // Re-exports for public API
 pub use color::*;
-pub use ports::{HardwarePorts, PortWriter, VgaPortOps};
+pub use crate::hardware::ports::{HardwarePorts, PortWriter, VgaPortOps};
 // VGA graphics modes
 pub use setup::{
     detect_and_init_vga_graphics, detect_cirrus_vga, init_vga_graphics, init_vga_text_mode,
