+++ petroleum/src/graphics/setup.rs
@@ -1,4 +1,5 @@
-use super::ports::{HardwarePorts, PortWriter, VgaPortOps};
+use crate::hardware::ports::{HardwarePorts, PortWriter, VgaPortOps};
+use crate::{write_port_sequence, write_vga_register};
 use super::registers::{
     ATTRIBUTE_CONFIG, ATTRIBUTE_TEXT_CONFIG, CRTC_CONFIG, CRTC_TEXT_CONFIG, GRAPHICS_CONFIG,
     GRAPHICS_TEXT_CONFIG, SEQUENCER_CONFIG, SEQUENCER_TEXT_CONFIG,
@@ -133,7 +134,7 @@ pub fn setup_vga_text_mode() {
     setup_misc_output();
 
     // Unlock CRTC registers
-    let crtc_unlock = super::ports::RegisterConfig {
+    let crtc_unlock = crate::hardware::ports::RegisterConfig {
         index: 0x11,
         value: 0x0E,
     };
